PNG image is now served!!!
Excited about this:
1. requests to the /db_image endpoint successfully get the png
out of the database and return back that image.

2. requests to /:questionId endpoint successfully get the png
of the question in the database matching that ObjectId and
spit back the image.

A big learning for me was the fact that the image is stored
as a Binary type in the database. I was trying to call img.length
and set that value in the header, but in fact img.length is a function
so I was getting a weird crashing message.

Another big item was that my Python parsing script on my PC was
calling f.open(filename, 'r') and so calls to .read() were broken
by newlines. The fix was f.open(filename, 'rb')